% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_centroid.R
\name{calc_centroid}
\alias{calc_centroid}
\title{Calculate the centroid, centroidal distance, and area Q}
\usage{
calc_centroid(df_input_response)
}
\arguments{
\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}
}
\value{
A `list` with the following components:
  \describe{
    \item{centroid}{A numeric vector of length 2, giving the x and y coordinates of the centroid.}
    \item{distance}{The centroidal distance (numeric).}
    \item{moment}{The area moment (numeric), often referred to as area Q.}
  }
}
\description{
Calculate the centroid, centroidal distance, and area Q
}
\examples{
data("df_input_response_example")
calc_centroid(df_input_response = df_input_response_example)

}
