% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_chance_corrected.R
\name{calc_chance_corrected}
\alias{calc_chance_corrected}
\title{Calculate the Chance-Corrected R for Language Assessment}
\usage{
calc_chance_corrected(df_input_response)
}
\arguments{
\item{df_input_response}{A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.}
}
\value{
A numeric vector containing the chance-corrected R value for the provided data. This value can help assess consistency in language evaluations by controlling for chance agreement.
}
\description{
This function computes the chance-corrected R value, a measure used in the VOX analysis to evaluate agreement while accounting for chance. The R value adjusts for potential randomness in scoring or response patterns, providing a more accurate assessment of inter-rater or inter-response agreement.
}
\examples{
library(dplyr)
data("df_input_response_example")
calc_chance_corrected(df_input_response = df_input_response_example)

}
\references{
For more information on chance-corrected measures in agreement assessments, see:
\url{https://peterstatistics.com/CrashCourse/5-ThreeVarPair/binary/MultipleBinaryPaired3c.html}
}
