% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_speakers_score.R
\name{calc_speakers_SCoRE}
\alias{calc_speakers_SCoRE}
\title{Calculate the Speaker's SCoRE}
\usage{
calc_speakers_SCoRE(df_summarized_response)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `util_summarize_response` for how to produce.}
}
\value{
A vector with the Speaker's SCoRE
}
\description{
Produces a vector with the Speaker's SCoRE
}
\examples{
data("df_summarized_response_example")
# Must be limited to one date.
dat <- df_summarized_response_example \%>\%
  filter(date_of_evaluation == max(date_of_evaluation))
calc_speakers_SCoRE(df_summarized_response = dat)
}
