% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_input_response_example}
\alias{df_input_response_example}
\title{Speaker Response Inputs (Example Data)}
\format{
A data frame with 7 rows and 6 variables:
\describe{
  \item{date_of_evaluation}{The date on which the evaluation took place.}
  \item{referent}{The name of the referent associated with the response, representing the object or entity involved in the verbal episode.}
  \item{conversing}{Binary indicator showing if the speaker provided a conversational response to the referent (1 = Yes, 0 = No).}
  \item{labeling}{Binary indicator showing if the speaker could label the referent (1 = Yes, 0 = No).}
  \item{echoing}{Binary indicator showing if the speaker could repeat the referent's name (1 = Yes, 0 = No).}
  \item{requesting}{Binary indicator showing if the speaker could request the referent's name (1 = Yes, 0 = No).}
}
}
\usage{
data(df_input_response_example)
}
\description{
This data set provides an example structure for `df_input_response`, a core data frame in the VOX Analysis process used to record speaker responses during language assessments. Each entry is a binary indicator (1 or 0) that denotes whether the speaker responded to a given referent across different verbal operants.
}
\seealso{
[df_input_response_previous_example] for past evaluation examples.
}
\keyword{datasets}
