% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_evaluation_date_entry_doc.R,
%   R/mod_evaluation_date_entry_server.R, R/mod_evaluation_date_entry_ui.R
\name{mod_evaluation_date_entry}
\alias{mod_evaluation_date_entry}
\alias{mod_evaluation_date_entry_server}
\alias{mod_evaluation_date_entry_ui}
\title{Evaluation Date Form for Appended Entries}
\usage{
mod_evaluation_date_entry_server(
  id,
  df_input_speaker_info,
  df_input_response_previous
)

mod_evaluation_date_entry_ui(id)
}
\arguments{
\item{id}{The namespace `id`, used to uniquely identify Shiny modules and UI components within the VOX Analysis package.}

\item{df_input_speaker_info}{A `data.frame` containing general information about a speaker (the subject of the VOX Analysis), as provided by the user. For an example, run `data("df_input_speaker_info")`.}

\item{df_input_response_previous}{A `data.frame` similar to `df_summarized_response` but represents past evaluations uploaded by the end user to distinguish prior assessments from current ones.}
}
\description{
This Shiny module provides a user interface and server function that allows the user to specify their current evaluation date when they are appending new records to an existing data set.
}
