% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_response_entry_doc.R,
%   R/mod_response_entry_server.R, R/mod_response_entry_ui.R
\name{mod_response_entry}
\alias{mod_response_entry}
\alias{mod_response_entry_server}
\alias{mod_response_entry_ui}
\title{A Shiny module to enter speaker data and render a report}
\usage{
mod_response_entry_server(
  id,
  df_input_speaker_info,
  df_input_response_previous = NULL
)

mod_response_entry_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_speaker_info}{A \code{data.frame} that contains general info about a speaker, provided by the user. Run \code{data("df_input_speaker_info")} to produce an example.}
}
\description{
A module that generates a UI and server function to allow the user to enter speaker data and render a report.
}
\examples{
if (interactive()) {
data("df_input_speaker_info_example")

ui <- page(
  useShinyjs(),
  mod_response_entry_ui("data_entry")
)

server <- function(input, output, session) {
  mod_response_entry_server("data_entry", df_input_speaker_info_example)
}

}
shinyApp(ui, server)

if (interactive()) {
### Alternative
run_example_measurment_entry_app()
}

}
