% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_primary_doc.R,
%   R/mod_results_primary_server.R, R/mod_results_primary_ui.R
\name{mod_results_primary}
\alias{mod_results_primary}
\alias{mod_results_primary_server}
\alias{mod_results_primary_ui}
\title{A Shiny module to generate a summary report based on the response inputs}
\usage{
mod_results_primary_server(id, df_input_speaker_info, df_input_response)

mod_results_primary_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_speaker_info}{A `data.frame` that contains general info about a speaker, provided by the user. Run `data("df_input_speaker_info")` to produce an example.}

\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}
}
\description{
A module that generates a UI and server function to generate a summary report.

Can operate independently of other modules.
}
\examples{
data("df_input_speaker_info_example")
data("df_input_response_example")

ui <- page(
  mod_results_primary_ui("data_input")
)

server <- function(input, output, session) {

  mod_results_primary_server(
    "data_input",
    df_input_speaker_info = df_input_speaker_info_example,
    df_input_response = df_input_response_example)

}

shinyApp(ui, server)

### Alternative
run_example_results_app()

}
