% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_section_detailed_analysis_doc.R,
%   R/mod_results_section_detailed_analysis_server.R,
%   R/mod_results_section_detailed_analysis_ui.R
\name{mod_results_section_detailed_analysis}
\alias{mod_results_section_detailed_analysis}
\alias{mod_results_section_detailed_analysis_server}
\alias{mod_results_section_detailed_analysis_ui}
\title{Detailed Analysis Section Module}
\usage{
mod_results_section_detailed_analysis_server(id, df_input_response)

mod_results_section_detailed_analysis_ui(id)
}
\arguments{
\item{id}{The namespace `id`, used to uniquely identify Shiny modules and UI components within the VOX Analysis package.}

\item{df_input_response}{A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.}
}
\description{
This Shiny module generates a detailed analysis section for VOX Analysis, designed to be used within the `mod_results_primary` module. It provides a comprehensive view of the speaker's responses by generating multiple components:
\itemize{
  \item \strong{Vox pie chart}: Visualizes the distribution of different response types.
  \item \strong{Vox line chart}: Shows trends in responses over time.
  \item \strong{Cochran's Q test}: Performs a statistical test to assess consistency across responses.
  \item \strong{Prompt hierarchy table}: Displays a structured view of response prompts in a hierarchy.
}
}
\examples{
if (interactive()) {
  # Load example data
  data("df_input_response_example")

  # Define the UI with the detailed analysis section
  ui <- page(
    mod_results_section_detailed_analysis_ui("detailed_analysis")
  )

  # Set up the server to include detailed analysis functionality
  server <- function(input, output, session) {
    mod_results_section_detailed_analysis_server(
      "detailed_analysis",
      df_input_response = df_input_response_example
    )
  }

  # Run the Shiny app
  shinyApp(ui, server)
}

}
