% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_speaker_data_entry_doc.R,
%   R/mod_speaker_data_entry_server.R, R/mod_speaker_data_entry_ui.R
\name{mod_speaker_data_entry}
\alias{mod_speaker_data_entry}
\alias{mod_speaker_data_entry_server}
\alias{mod_speaker_data_entry_ui}
\title{Speaker Data Entry Module}
\usage{
mod_speaker_data_entry_server(id)

mod_speaker_data_entry_ui(id)
}
\arguments{
\item{id}{The namespace `id`, used to uniquely identify Shiny modules and UI components within the VOX Analysis package.}
}
\description{
This Shiny module provides a user interface for entering speaker data, triggering a data entry prompt followed by a report output. It can be used to collect speaker information for VOX Analysis.
}
\examples{
if (interactive()) {
  # Basic usage example
  ui <- page(
    useShinyjs(),
    mod_speaker_data_entry_ui("speaker_entry")
  )

  server <- function(input, output, session) {
    mod_speaker_data_entry_server("speaker_entry")
  }

  shinyApp(ui, server)
}
}
