% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_speaker_data_info_doc.R,
%   R/mod_speaker_data_info_server.R, R/mod_speaker_data_info_ui.R
\name{mod_speaker_data_info}
\alias{mod_speaker_data_info}
\alias{mod_speaker_data_info_server}
\alias{mod_speaker_data_info_ui}
\title{A module to view and modify the speaker data}
\usage{
mod_speaker_data_info_server(id, df_input_speaker_info)

mod_speaker_data_info_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_speaker_info}{A `data.frame` that contains general info about a speaker, provided by the user. Run `data("df_input_speaker_info")` to produce an example.}
}
\description{
A module that allows the user to view and modify speaker data.
}
\examples{
data("df_input_speaker_info_example")
ui <- page(
   useShinyjs(),
   mod_speaker_data_info_ui("speaker")
)

server <- function(input, output, session) {
   mod_speaker_data_info_server("speaker", df_input_speaker_info_example)
}
shinyApp(ui, server)

}
