% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_starting_page_doc.R,
%   R/mod_starting_page_server.R, R/mod_starting_page_ui.R
\name{mod_starting_page}
\alias{mod_starting_page}
\alias{mod_starting_page_server}
\alias{mod_starting_page_ui}
\title{Application Home Page Module}
\usage{
mod_starting_page_server(id)

mod_starting_page_ui(id)
}
\arguments{
\item{id}{The namespace `id`, used to uniquely identify Shiny modules and UI components within the VOX Analysis package.}
}
\description{
This Shiny module defines the home page for the VOX Analysis application, serving as the main entry point for users. It contains the primary layout and navigation elements for accessing various sections of the app.
}
\examples{
if (interactive()) {
  # Define the UI with the application's home page
  ui <- page(
    useShinyjs(),
    mod_starting_page_ui("home_page")
  )

  # Set up the server for the home page module
  server <- function(input, output, session) {
    mod_starting_page_server("home_page")
  }

  # Run the Shiny app
  shinyApp(ui, server)
}

}
