% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_map.R
\docType{data}
\name{plot_color_mapping}
\alias{plot_color_mapping}
\title{Color map for charts}
\format{
An object of class \code{character} of length 4.
}
\usage{
plot_color_mapping
}
\description{
Produces hexadecimal color codes for plotting.
}
\examples{
plot_color_mapping


}
\keyword{datasets}
