% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_measurements.R
\name{summarize_response}
\alias{summarize_response}
\title{Group response}
\usage{
summarize_response(df_input_response)
}
\arguments{
\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}
}
\description{
A utility function to sum the response by the type of response.
}
\examples{
summarize_response(
  df_input_response = df_input_response_example
  )
}
