% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_cochrans_q_test.R
\name{table_cochrans_q_test}
\alias{table_cochrans_q_test}
\title{Create summary table with cochran's q test and supporting tests}
\usage{
table_cochrans_q_test(df_input_response)
}
\arguments{
\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}
}
\description{
Produces data.frame with cochran's q test, chance-corrected R, and p-value.
}
\examples{
data("df_input_response_example")
table_cochrans_q_test(df_input_response = df_input_response_example)


}
