% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_custom_file_upload.R
\name{tag_custom_file_upload}
\alias{tag_custom_file_upload}
\title{Custom File Upload Control}
\usage{
tag_custom_file_upload(
  inputId,
  multiple = FALSE,
  accept = NULL,
  buttonLabel = "Browse...",
  placeholder = "No file selected",
  ...
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{multiple}{Whether the user should be allowed to select and upload
multiple files at once. \strong{Does not work on older browsers, including
Internet Explorer 9 and earlier.}}

\item{accept}{A character vector of "unique file type specifiers" which gives
the browser a hint as to the type of file the server expects. Many browsers
use this prevent the user from selecting an invalid file.

A unique file type specifier can be:
\itemize{
\item A case insensitive extension like \code{.csv} or \code{.rds}.
\item A valid MIME type, like \code{text/plain} or \code{application/pdf}
\item One of \verb{audio/*}, \verb{video/*}, or \verb{image/*} meaning any audio, video,
or image type, respectively.
}}

\item{buttonLabel}{The label used on the button. Can be text or an HTML tag
object.}

\item{placeholder}{The text to show before a file has been uploaded.}

\item{...}{
  Arguments passed on to \code{\link[shiny:fileInput]{shiny::fileInput}}
  \describe{
    \item{\code{label}}{Display label for the control, or \code{NULL} for no label.}
    \item{\code{width}}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
    \item{\code{capture}}{What source to use for capturing image, audio or video data.
This attribute facilitates user access to a device's media capture
mechanism, such as a camera, or microphone, from within a file upload
control.

A value of \code{user} indicates that the user-facing camera and/or microphone
should be used. A value of \code{environment} specifies that the outward-facing
camera and/or microphone should be used.

By default on most phones, this will accept still photos or video. For
still photos only, also use \code{accept="image/*"}. For video only, use
\code{accept="video/*"}.}
  }}
}
\value{
A UI element for file upload, similar to `shiny::fileInput`.
}
\description{
Create a file upload control that can be used to upload one or more files.
}
\details{
This is modified from `shiny::fileInput` to make the format more consistent with the VOX Analysis user-interface.
}
