% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_custom_numeric_input.R
\name{tag_custom_input}
\alias{tag_custom_input}
\title{A custom numeric input for the language assessment app}
\usage{
tag_custom_input(
  input_id,
  value = 0,
  align_center = TRUE,
  placeholder = "Enter value"
)
}
\arguments{
\item{input_id}{An id}

\item{value}{A numeric value}
}
\description{
This function is used in `tag_table_response_inputs` to produce a matrix-like data entry input.
}
