% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_div_ones_zeros.R
\name{tag_div_ones_zeros}
\alias{tag_div_ones_zeros}
\title{A `taglist` containing warning message regarding all 1s and 0s for responses}
\usage{
tag_div_ones_zeros(output_name = "This measure")
}
\arguments{
\item{output_name}{Name of the output (i.e., Cochran's Q, Area Q)}
}
\value{
A
}
\description{
This `taglist` provides a warning message to inform the user that a measure or data visualizations can't be produced because only 1s and 0s were entered.
}
\seealso{
[util_check_for_ones_zeros] for generating the input data frame that defines the prompt hierarchy.
}
