% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_spinner.R
\name{tag_spinner}
\alias{tag_spinner}
\title{Add a Spinner to a UI Element}
\usage{
tag_spinner(tag_to_pass, color = "#cce5f3")
}
\arguments{
\item{tag_to_pass}{The Shiny UI element to wrap with a spinner. Required.}

\item{color}{A character string specifying the spinner color in hexadecimal format. Default is `"#cce5f3"`.}
}
\value{
A UI element wrapped with a spinner.
}
\description{
Wraps a Shiny UI element with a spinner to indicate loading or processing activity.
}
\examples{
\dontrun{
tag_spinner(shiny::textOutput("output_id"))
}
}
