% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table_prompt_hierarchy.R
\name{tag_table_prompt_hierarchy}
\alias{tag_table_prompt_hierarchy}
\title{Custom HTML Table to Report Prompt Hierarchy}
\usage{
tag_table_prompt_hierarchy(table_prompt_hierarchy)
}
\arguments{
\item{table_prompt_hierarchy}{A `data.frame` generated by `table_prompt_hierarchy`.}
}
\value{
An HTML table displaying the prompt hierarchy in a static format for reporting purposes.
}
\description{
This function creates a custom HTML table to display the prompt hierarchy on the results page of the VOX Analysis application. It provides a structured, read-only view of the hierarchy, which is generated based on the intervention steps for the speaker.
}
\seealso{
[table_prompt_hierarchy] for generating the input data frame that defines the prompt hierarchy.
}
