% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table_measurement_results.R
\name{tag_table_response_results}
\alias{tag_table_response_results}
\title{A custom html table to reports response}
\usage{
tag_table_response_results(df_input_response)
}
\arguments{
\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}
}
\description{
A custom html table that reports the response.

Unlike \code{tag_table_response_input}, this cannot be interacted with by the user in a Shiny session and is a static rendering.
}
