% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_data_upload.R
\name{util_check_data_upload}
\alias{util_check_data_upload}
\title{Quality check df_input_response}
\usage{
util_check_data_upload(df_input_response)
}
\arguments{
\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}
}
\description{
A helpful function to ensure that `df_input_response` follows the right data model.

This is particular useful with quality checking data the user uploaded via `mod_upload_data_set`.

Returns `Good` or `Bad` to avoid ambiguity.
}
\examples{
data("df_input_response_example")
util_check_data_upload(df_input_response = df_input_response_example)


}
