% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_check_for_ones_zeros.R
\name{util_check_for_ones_zeros}
\alias{util_check_for_ones_zeros}
\title{Check if `df_input_response` is all 1s or 0s}
\usage{
util_check_for_ones_zeros(df_input_response, ind_check_only_for_zero = FALSE)
}
\arguments{
\item{df_input_response}{A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.}

\item{ind_check_only_for_zero}{Only evaluate if they are all zeros}
}
\description{
This function performs a quality check to determine whether `df_input_response` has only 1s or 0s as entries.
}
\examples{
# Load example data, which has both 1s and 0s
data("df_input_response_example")

# This will return FALSE
util_check_for_ones_zeros(df_input_response = df_input_response_example)

# Create data frame with all 0s
df_zeroed <- data.frame(
  date_of_evaluation = rep('2024-08-23', 6),
  referent = c('Apple', 'Ball', 'Tablet', 'Pencil', 'Army Man', 'Mouse'),
  conversing = rep(0, 6),
  labeling = rep(0, 6),
  echoing = rep(0, 6),
  requesting = rep(0, 6)
 )

# This will return TRUE
util_check_for_ones_zeros(df_input_response = df_zeroed)

# Create data frame with all 1s
df_ones <- data.frame(
  date_of_evaluation = rep('2024-08-23', 6),
  referent = c('Apple', 'Ball', 'Tablet', 'Pencil', 'Army Man', 'Mouse'),
  conversing = rep(1, 6),
  labeling = rep(1, 6),
  echoing = rep(1, 6),
  requesting = rep(1, 6)
 )

# This will return TRUE
util_check_for_ones_zeros(df_input_response = df_ones)


}
