% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_download_handler.R
\name{util_download_handler}
\alias{util_download_handler}
\alias{util_download_handler_word}
\alias{util_download_handler_dataset}
\title{Download handlers for producing .csv and Word exports}
\usage{
util_download_handler_word(
  df_input_speaker_info,
  df_input_response,
  df_summarized_response
)

util_download_handler_dataset(df_input_speaker_info, df_input_response)
}
\arguments{
\item{df_input_speaker_info}{A \code{data.frame} that contains general info about a speaker, provided by the user. Run \code{data("df_input_speaker_info")} to produce an example.}

\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}

\item{df_summarized_response}{A \code{data.frame} that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See \code{summarize_response} for how to produce.}
}
\description{
\code{util_download_handler_word} and \code{util_download_handler_dataset} pass params on to \code{downloadHandler}.

\code{util_download_handler_word} generates a Word version of the results page.

\code{util_download_handler_dataset} generates a \code{.csv} with the inputs provided by the user.
}
