% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shiny_remove_and_hide_flex.R
\name{util_shiny_remove_and_hide_flex}
\alias{util_shiny_remove_and_hide_flex}
\title{Removes flex classes from interfering with hide function}
\usage{
util_shiny_remove_and_hide_flex(inputId)
}
\arguments{
\item{inputId}{The tag ID}
}
\description{
Whenever a class of \verb{d-flex flex-fill} is used in an HTML tag \code{class}, \code{shinyjs::hide()} does not work.

This utility function removes the classes and hides the tag.
}
