% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_chance_corrected.R
\name{calc_chance_corrected}
\alias{calc_chance_corrected}
\title{Provided the chance corrected R calculation}
\usage{
calc_chance_corrected(df_input_response)
}
\arguments{
\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}
}
\value{
A `vector`.
}
\description{
Calculate the chance corrected R calculation for the language assessment.
}
\examples{
library(dplyr)
data("df_input_response_example")
calc_chance_corrected(df_input_response = df_input_response_example)

}
\references{
\url{https://peterstatistics.com/CrashCourse/5-ThreeVarPair/binary/MultipleBinaryPaired3c.html}
}
