% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cochran_qtest.R
\name{calc_cochran_qtest}
\alias{calc_cochran_qtest}
\title{Perform Cochran's Q Test for Multiple Response Comparisons}
\usage{
calc_cochran_qtest(formula, data, alpha = 0.05, p.method = "bonferroni")
}
\arguments{
\item{formula}{A formula specifying the model to use in the test. Typically, the format is `value ~ name | referent`, where `value` is the binary response variable and `name` represents different conditions or responses.}

\item{data}{A `data.frame` containing the binary responses for each verbal operant (e.g., `Conversing`, `Labeling`, `Echoing`, and `Requesting`). The data should be transformed with `pivot_longer` so that the binary operant names are in a `name` column, and responses are in a `value` column.}

\item{alpha}{Significance level for the test (e.g., 0.05).}

\item{p.method}{Method for p-value adjustment for multiple comparisons; VOX analysis uses `bonferroni`.}
}
\value{
A `list` containing the following components:
  \describe{
    \item{statistic}{The test statistic for Cochran's Q.}
    \item{p.value}{The p-value associated with the test statistic, after adjustment if `p.method` is specified.}
    \item{...}{Other values are returned, but not used for the purpose of the VOX Analysis app.}
  }
}
\description{
This function performs Cochran's Q Test, a non-parametric test for examining differences in proportions across multiple related groups. It is based on the `RVAideMemoire` package's implementation, adapted here to avoid external dependencies and ensure seamless integration. Cochran's Q Test is commonly used in repeated measures designs or studies with binary response data.
}
\examples{
# Example of Cochran's Q Test using sample response data
library(dplyr)
library(tidyr)
data("df_input_response_example")
 dat <- df_input_response_example \%>\%
    select("referent", "conversing", "labeling", "echoing", "requesting") \%>\%
   pivot_longer(
     cols = c("conversing", "labeling", "echoing", "requesting"),
     names_to = "name",
     values_to = "value") \%>\%
     arrange(.data$name)
calc_cochran_qtest(
  formula = value ~ name | referent,
  data = dat,
  alpha = 0.05,
  p.method = "bonferroni"
)

}
