% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cochran_qtest.R
\name{calc_cochran_qtest}
\alias{calc_cochran_qtest}
\title{Cochran's Q test}
\usage{
calc_cochran_qtest(formula, data, alpha = 0.05, p.method = "fdr")
}
\arguments{
\item{formula}{A formula to pass through. The app currently uses \code{value ~ name | referent}.}

\item{data}{A \code{data.frame} containing \code{Conversing}, \code{Labeling}, \code{Echoing}, and \code{Requesting}. Must have \code{pivot_longer} applied, with \code{name} and \code{value} being the two new columns.}

\item{alpha}{alpha level.}

\item{p.method}{p method. Recommend "fdr".}
}
\value{
A \code{list}.
}
\description{
This is heavily based on Cochran's Q Test from \code{RVAideMemoire} package.

There were several dependency issues with the package, so it was re-written to be a built-in function for this package with some modification.
}
