% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_speakers_score.R
\name{calc_speakers_SCoRE}
\alias{calc_speakers_SCoRE}
\title{Calculate the speaker's SCoRE}
\usage{
calc_speakers_SCoRE(df_summarized_response)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `util_summarize_response` for how to produce.}
}
\description{
Produces a data.frame with the prompt hierarchy
}
\examples{
data("df_summarized_response_example")
 ## Function only works with one date at a time
dat <- df_summarized_response_example \%>\%
    filter(date_of_evaluation == max(date_of_evaluation))
calc_speakers_SCoRE(df_summarized_response = dat)


}
