% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common-params.R
\name{common-params}
\alias{common-params}
\title{Common parameters}
\arguments{
\item{id}{The namespace id.}

\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}

\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `util_summarize_response` for how to produce.}

\item{df_input_speaker_info}{A `data.frame` that contains general info about a speaker, provided by the user. Run `data("df_input_speaker_info")` to produce an example.}
}
\description{
Common parameters found throughout the language assessment package.
}
