% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_input_response_example}
\alias{df_input_response_example}
\title{Speaker Response Inputs (Example)}
\format{
A data frame with 7 rows and 6 variables:
\describe{
  \item{date_of_evaluation}{The date of the evaluation.}
  \item{referent}{Name of the referent associated with the response.}
  \item{conversing}{Indicates whether the speaker provided a thematic response to an object (1 = Yes, 0 = No).}
  \item{labeling}{Indicates whether the speaker could label an object (1 = Yes, 0 = No).}
  \item{echoing}{Indicates whether the speaker could repeat an object name (1 = Yes, 0 = No).}
  \item{requesting}{Indicates whether the speaker could request an object name (1 = Yes, 0 = No).}
}
}
\usage{
data(df_input_response_example)
}
\description{
This data set shows the structure expected for `df_input_response`, which is a common data set that's generated by the language assessment application.
}
\details{
Each entry is a 1 or 0, which indicates whether a speaker responded to the referent.
}
\seealso{
[df_input_response_previous_example]
}
\keyword{datasets}
