% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_output_exportable_example}
\alias{df_output_exportable_example}
\title{Export Data Structure (Example)}
\format{
A data frame with 12 columns:
\describe{
  \item{first_name}{The speaker's first name.}
  \item{last_name}{The speaker's last name.}
  \item{date_of_birth}{The speaker's date of birth.}
  \item{language_spoken}{The language spoken by the speaker during the evaluation.}
  \item{gender}{The speaker's gender.}
  \item{date_of_evaluation}{The date on which the evaluation was conducted.}
  \item{referent}{The name of the referent associated with the response, representing the object or entity involved in the verbal episode.}
  \item{referent_order}{The order the referent appeared in the session. Required for data visualizations and calculations regarding specific verbal episodes.}
  \item{conversing}{Binary indicator showing if the speaker provided a conversational response to the referent (1 = Yes, 0 = No).}
  \item{labeling}{Binary indicator showing if the speaker could label the referent (1 = Yes, 0 = No).}
  \item{echoing}{Binary indicator showing if the speaker could repeat the referent's name (1 = Yes, 0 = No).}
  \item{requesting}{Binary indicator showing if the speaker could request the referent's name (1 = Yes, 0 = No).}
}
}
\usage{
data(df_output_exportable_example)
}
\description{
This data set represents the structure required for exporting evaluation data. Any data re-imported into the application to append new records must follow this format to ensure compatibility and accurate processing.
}
\seealso{
See [df_input_response_example] for an example of individual response input data.
}
\keyword{datasets}
