% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_summarized_response_example}
\alias{df_summarized_response_example}
\title{Summarized response}
\format{
\describe{
 \item{date_of_evaluation}
 \item{conversing}
 \item{labeling}
 \item{echoing}
 \item{requesting}
}
}
\usage{
data(df_summarized_response_example)
}
\description{
This data set is a summation of the four response (conversing, labeling, echoing, and requesting). It is often used as an input for several plot functions.
}
\details{
It can be produced using `util_summarize_response`. (See `?util_summarize_meausurments`.)
}
\keyword{datasets}
