% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_summarized_response_example}
\alias{df_summarized_response_example}
\title{Summarized Response Data (Example)}
\format{
A data frame with 3 rows and 5 variables:
\describe{
  \item{date_of_evaluation}{The date on which the evaluation was conducted.}
  \item{conversing}{The total count of positive conversing responses for the given evaluation date.}
  \item{labeling}{The total count of positive labeling responses for the given evaluation date.}
  \item{echoing}{The total count of positive echoing responses for the given evaluation date.}
  \item{requesting}{The total count of positive requesting responses for the given evaluation date.}
}
}
\usage{
data(df_summarized_response_example)
}
\description{
This data set provides an example of the summarized response data, where the total occurrences of each response type (Conversing, Labeling, Echoing, and Requesting) are calculated for each evaluation date. This summary is often used as an input for plot functions and other analyses to assess patterns in response types over time.

The summarized data can be generated using the `util_summarize_response` function. See `?util_summarize_response` for more details.
}
\seealso{
See [util_summarize_response] for generating this summary data frame from raw response inputs.
}
\keyword{datasets}
