% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_bar_doc.R,
%   R/mod_download_bar_server.R, R/mod_download_bar_ui.R
\name{mod_download_bar}
\alias{mod_download_bar}
\alias{mod_download_bar_server}
\alias{mod_download_bar_ui}
\title{A module for a download bar}
\usage{
mod_download_bar_server(
  id,
  df_input_speaker_info,
  df_input_response,
  df_summarized_response
)

mod_download_bar_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_speaker_info}{A \code{data.frame} that contains general info about a speaker, provided by the user. Run \code{data("df_input_speaker_info")} to produce an example.}

\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}

\item{df_summarized_response}{A \code{data.frame} that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See \code{summarize_response} for how to produce.}
}
\description{
A Shiny module to allow a user to download the exportable data set and word doc report.
}
\examples{
if (interactive()) {
data("df_input_speaker_info_example")
data("df_input_response_example")
data("df_summarized_response_example")

ui <- page(
  mod_download_bar_ui("download_bar")
)

server <- function(input, output, session) {

  mod_download_bar_server(
    "download_bar",
    df_input_speaker_info = df_input_speaker_info_example,
    df_input_response = df_input_response_example,
    df_summarized_response = df_summarized_response_example)

}

shinyApp(ui, server)
}

}
