% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_section_area_q_doc.R,
%   R/mod_results_section_area_q_server.R, R/mod_results_section_area_q_ui.R
\name{mod_results_section_area_q}
\alias{mod_results_section_area_q}
\alias{mod_results_section_area_q_server}
\alias{mod_results_section_area_q_ui}
\title{A module for an Area Q summary section}
\usage{
mod_results_section_area_q_server(id, df_input_response, v_primary_date)

mod_results_section_area_q_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}
}
\description{
A Shiny module that generates an Area Q summary section, intended to be used in \code{mod_results_primary}.
}
\examples{

if (interactive()) {
data("df_input_response_example")

ui <- page(
  mod_results_section_area_q_ui("section_area_q")
)

server <- function(input, output, session) {

  mod_results_section_area_q_server(
    "section_area_q",
    df_input_response_example,
    v_primary_date = max(df_input_response_example$date_of_evaluation))

}

shinyApp(ui, server)
}

}
