% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_section_detailed_analysis_doc.R,
%   R/mod_results_section_detailed_analysis_server.R,
%   R/mod_results_section_detailed_analysis_ui.R
\name{mod_results_section_detailed_analysis}
\alias{mod_results_section_detailed_analysis}
\alias{mod_results_section_detailed_analysis_server}
\alias{mod_results_section_detailed_analysis_ui}
\title{A module for a  detailed analysis section}
\usage{
mod_results_section_detailed_analysis_server(id, df_input_response)

mod_results_section_detailed_analysis_ui(id)
}
\arguments{
\item{id}{The namespace id.}

\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}
}
\description{
A Shiny module to generate a detailed analysis section, intended to be used in `mod_results_primary`.

This module will generate Vox pie chart, Vox line chart, Cochran's Q test, and prompt hierarchy table.
}
\examples{

if (interactive()) {
data("df_input_speaker_info_example")
data("df_input_response_example")
data("df_input_response_previous_example")

ui <- page(
  mod_results_section_detailed_analysis_ui("detailed_analysis")
)

server <- function(input, output, session) {

  mod_results_section_detailed_analysis_server(
    "detailed_analysis",
    df_input_speaker_info = df_input_speaker_info_example,
    df_input_response = df_input_response_example,
    df_input_response_previous = df_input_response_previous_example)

}

shinyApp(ui, server)
}


}
