% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_speaker_data_entry_doc.R,
%   R/mod_speaker_data_entry_server.R, R/mod_speaker_data_entry_ui.R
\name{mod_speaker_data_entry}
\alias{mod_speaker_data_entry}
\alias{mod_speaker_data_entry_server}
\alias{mod_speaker_data_entry_ui}
\title{A Shiny module to input speaker data}
\usage{
mod_speaker_data_entry_server(id)

mod_speaker_data_entry_ui(id)
}
\arguments{
\item{id}{The namespace id.}
}
\description{
A module that allows the user to enter speaker data via a prompt.

It subsequently triggers a data entry prompt and a report output.
}
\examples{
ui <- page(
  useShinyjs(),
  mod_speaker_data_entry_ui("speaker_entry")
)

server <- function(input, output, session) {
  mod_speaker_data_entry_server("speaker_entry")
}

shinyApp(ui, server)

### Alternative: include previous speaker data
data("df_input_speaker_info_example")

ui <- page(
  useShinyjs(),
  mod_speaker_data_entry_ui("speaker_entry")
)

server <- function(input, output, session) {
  mod_speaker_data_entry_server(
  "speaker_entry",
  df_input_speaker_info = df_input_speaker_info_example)
}

shinyApp(ui, server)

}
