% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_starting_page_doc.R,
%   R/mod_starting_page_server.R, R/mod_starting_page_ui.R
\name{mod_starting_page}
\alias{mod_starting_page}
\alias{mod_starting_page_server}
\alias{mod_starting_page_ui}
\title{A module for the application's home page}
\usage{
mod_starting_page_server(id)

mod_starting_page_ui(id)
}
\arguments{
\item{id}{The namespace id.}
}
\description{
A Shiny module containing the home page for the application (basically the entire app).
}
\examples{
ui <- page(
   useShinyjs(),
   mod_starting_page_ui("home_page")
)

server <- function(input, output, session) {
   mod_starting_page_server("home_page")
}
shinyApp(ui, server)

}
