% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area_q.R
\name{plot_area_q}
\alias{plot_area_q}
\title{Area Q Plot}
\usage{
plot_area_q(df_summarized_response, v_primary_date)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `summarize_response` for how to produce.}
}
\description{
Produces an area Q plot over time for the data set provided.
}
\examples{
data("df_summarized_response_example")
plot_area_q(df_summarized_response = df_summarized_response_example)

}
