% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_color_map.R
\docType{data}
\name{plot_color_mapping}
\alias{plot_color_mapping}
\title{Color Map for Response Type Charts}
\format{
An object of class \code{character} of length 4.
}
\usage{
plot_color_mapping
}
\description{
A predefined set of hexadecimal color codes for use in VOX Analysis charts, mapped to specific response types (Conversing, Labeling, Echoing, Requesting). This color map ensures consistent and visually distinguishable colors across charts, enhancing interpretability.
}
\keyword{datasets}
