% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vox_line.R
\name{plot_vox_line}
\alias{plot_vox_line}
\title{Plot a Vox Line Chart}
\usage{
plot_vox_line(df_input_response, hide_heading = TRUE, doc_version_ind = FALSE)
}
\arguments{
\item{df_input_response}{A `data.frame` containing the referent names and their response. Run `data("df_input_response_example")` or `data("df_input_response_previous_example")` for example.}

\item{hide_heading}{A logical indicator to hide the heading. It defaults to TRUE.}
}
\description{
Produces a Vox line chart
}
\examples{
data("df_input_response_example")
plot_vox_line(df_input_response = df_input_response_example)


}
