% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vox_piechart.R
\name{plot_vox_piechart}
\alias{plot_vox_piechart}
\title{Plot a Vox Pie Chart}
\usage{
plot_vox_piechart(df_summarized_response, ind_doc_version = FALSE)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `util_summarize_response` for how to produce.}

\item{ind_doc_version}{`TRUE` or `FALSE` value to indicate whether a plot needs to be rendered for a Word doc. Passing `TRUE` will change various font sizes and other aesthetic parameters of the plot to make it more Word-friendly.}
}
\description{
Produces a Vox pie chart.
}
\examples{
library(dplyr)
data("df_summarized_response_example")
dat <- df_summarized_response_example \%>\%
    filter(date_of_evaluation == max(date_of_evaluation)) ## Plot only works with one date at a time
plot_vox_piechart(df_summarized_response = dat)


}
