% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vox_piechart.R
\name{plot_vox_piechart}
\alias{plot_vox_piechart}
\title{Plot a Vox Pie Chart}
\usage{
plot_vox_piechart(df_summarized_response)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `summarize_response` for how to produce.}
}
\description{
Produces a Vox pie chart.
}
\examples{
data("df_summarized_response_example")
dat <- df_summarized_response_example \%>\%
    filter(date_of_evaluation == max(date_of_evaluation)) ## Plot only works with one date at a time
plot_vox_piechart(df_summarized_response = dat)


}
