% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\alias{run_example_results_app}
\alias{run_example_response_entry_app}
\alias{run_example_upload_data_set_app}
\alias{run_example_speaker_data_entry_app}
\title{Helper functions to run and test various applications / modules}
\usage{
run_example_results_app()

run_example_response_entry_app()

run_example_upload_data_set_app(add_new_data_ind = TRUE)

run_example_speaker_data_entry_app()
}
\description{
A series of helper functions to allow easy demonstration of various application / module combos.

Example data is included within the functions.
}
\examples{
# Runs the results page
run_example_results_app()

# Runs the response input page
run_example_response_entry_app()

# Runs the upload data page. Change add_new_data_ind = TRUE to include response_entry page
run_example_upload_data_set_app(add_new_data_ind = FALSE)

}
