% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\alias{run_example_results_app}
\alias{run_example_response_entry_app}
\alias{run_example_upload_data_set_app}
\alias{run_example_speaker_data_entry_app}
\title{Helper Functions to Demonstrate Application and Module Combinations}
\usage{
run_example_results_app()

run_example_response_entry_app()

run_example_upload_data_set_app(ind_add_new_data = TRUE)

run_example_speaker_data_entry_app()
}
\arguments{
\item{ind_add_new_data}{Logical (`TRUE` or `FALSE`): Indicates whether the user intends to manually enter new data. Setting this to `TRUE` triggers the inclusion of the `mod_speaker_data_entry` module, enabling manual data entry by the user.}
}
\description{
This set of helper functions allows users to easily run and test various combinations of VOX Analysis applications and modules. Each function includes example data, providing a straightforward way to demonstrate different functionalities without requiring additional setup.
}
\examples{
# Run the results page application
if (interactive()) {
  run_example_results_app()
}

# Run the response input page application
if (interactive()) {
  run_example_response_entry_app()
}

# Run the upload data page application
# Set ind_add_new_data = TRUE to include the response entry page
if (interactive()) {
  run_example_upload_data_set_app(ind_add_new_data = FALSE)
}

# Run the speaker data entry application
if (interactive()) {
  run_example_speaker_data_entry_app()
}

}
