% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prompt_hierarchy.R
\name{table_prompt_hierarchy}
\alias{table_prompt_hierarchy}
\title{Prompt Hierachy Data Farme}
\usage{
table_prompt_hierarchy(df_summarized_response)
}
\arguments{
\item{df_summarized_response}{A `data.frame` that sums each verbal operant column (i.e., conversing, echoing, etc.) for a given evaluation date. See `util_summarize_response` for how to produce.}
}
\description{
Produces a data.frame with the prompt hierarchy
}
\examples{
data("df_summarized_response_example")
 ## Table only works with one date at a time
dat <- df_summarized_response_example \%>\%
    filter(date_of_evaluation == max(date_of_evaluation))
table_prompt_hierarchy(df_summarized_response = dat)


}
