% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_custom_numeric_input.R
\name{tag_custom_input}
\alias{tag_custom_input}
\title{Custom Numeric Input for Borderless Data Entry}
\usage{
tag_custom_input(
  input_id,
  value = 0,
  align_center = TRUE,
  placeholder = "Enter value"
)
}
\arguments{
\item{input_id}{A unique identifier for the input element.}

\item{value}{A numeric value to be displayed as the default input.}

\item{align_center}{Logical. If `TRUE`, centers the input alignment. Defaults to `FALSE`.}

\item{placeholder}{A character string to display as placeholder text when the input is empty.}
}
\value{
A custom, borderless numeric input UI element without numeric increment arrows.
}
\description{
This function creates a custom numeric input field, similar to `shiny::textInput`, but with a borderless design and no up / down arrows. It is used in `tag_table_response_inputs` to facilitate smooth, minimalistic numeric input in the VOX Analysis language assessment app.
}
