% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_kpi.R
\name{tag_kpi}
\alias{tag_kpi}
\title{Create a KPI taglist}
\usage{
tag_kpi(metric, header)
}
\arguments{
\item{metric}{A single value for the metric to report.}

\item{header}{A string value containing the header or KPI description.}
}
\description{
Produces a KPI taglist containing only a header and the KPI itself.
}
