% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table_prompt_hierarchy.R
\name{tag_table_prompt_hierarchy}
\alias{tag_table_prompt_hierarchy}
\title{A custom html table to report prompt hierarchy}
\usage{
tag_table_prompt_hierarchy(table_prompt_hierarchy)
}
\description{
A custom HTML table to report the prompt hierarchy on the results page of the application.
}
