% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table_measurement_input.R
\name{tag_table_response_input}
\alias{tag_table_response_input}
\title{A custom html table to allow the user to type in response}
\usage{
tag_table_response_input(ns_prefix, row_count)
}
\arguments{
\item{ns_prefix}{The namespace. Pass `session$ns(id)` to make the user inputs available in the session via `input`. This will ensure that the inputs are accessible in the same module.}

\item{row_count}{The number of rows to include}
}
\description{
This function generate an html table, which uses various tags and classes, as well as `tag_custom_input`, to allow for data entry.

Data that is collected during the session will be made available to on the server-side via `input`.
}
