% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_download_handler.R
\name{util_download_handler}
\alias{util_download_handler}
\alias{util_download_handler_dataset}
\title{Download Handlers for Generating .csv and Word Exports}
\usage{
util_download_handler_dataset(df_input_speaker_info, df_input_response)
}
\arguments{
\item{df_input_speaker_info}{A `data.frame` containing general information about a speaker (the subject of the VOX Analysis), as provided by the user. For an example, run `data("df_input_speaker_info")`.}

\item{df_input_response}{A `data.frame` containing referent names and their associated responses, typically recorded by the listener. For example data, run `data("df_input_response_example")` or `data("df_input_response_previous_example")`.}
}
\description{
These functions provide download handlers for exporting data in `.csv` and Word document formats:

\itemize{
  \item `util_download_handler_word`: Generates a Word version of the results page, formatted for reporting and easy sharing of VOX Analysis outputs.
  \item `util_download_handler_dataset`: Creates a `.csv` file containing the input data provided by the user, allowing for quick data export and further analysis.
}

Both functions pass parameters to `downloadHandler`, enabling customization of file output within the VOX Analysis app.
}
\seealso{
[util_generate_word_doc_report] for generating a Word document report based on VOX Analysis data.
}
