% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_generate_word_doc_report.R
\name{util_generate_word_doc_report}
\alias{util_generate_word_doc_report}
\title{Render Word Doc REport}
\usage{
util_generate_word_doc_report(
  df_input_speaker_info = NULL,
  df_input_response,
  file_name,
  temp_file = TRUE
)
}
\arguments{
\item{df_input_speaker_info}{A \code{data.frame} that contains general info about a speaker, provided by the user. Run \code{data("df_input_speaker_info")} to produce an example.}

\item{df_input_response}{A \code{data.frame} containing the referent names and their response. Run \code{data("df_input_response_example")} or \code{data("df_input_response_previous_example")} for example.}
}
\description{
Produces a word doc report using the \code{printable_report.qmd} found in the \code{inst} folder.
}
\examples{
data("df_input_speaker_info_example")
data("df_input_response_example")
util_generate_word_doc_report(
  df_input_speaker_info = df_input_speaker_info_example,
  df_input_response = df_input_response_example)


}
