% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shiny_remove_and_hide_flex.R
\name{util_shiny_remove_and_hide_flex}
\alias{util_shiny_remove_and_hide_flex}
\title{Remove Flex Classes for Compatibility with `shinyjs::hide()`}
\usage{
util_shiny_remove_and_hide_flex(inputId)
}
\arguments{
\item{inputId}{The ID of the HTML tag to modify and hide.}
}
\description{
When an HTML tag has the `d-flex flex-fill` classes applied, it can interfere with the functionality of `shinyjs::hide()`. This utility function removes these classes from the specified tag, ensuring `shinyjs::hide()` works as expected to hide the element.
}
